common DeleteButton component for repeating parts
